requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
filelock==3.17.0
flask==3.1.0
fsspec==2025.3.0
furl==2.1.4
grpcio==1.70.0
grpcio-tools==1.70.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
orderedmultidict==1.0.1
packaging==24.2
pillow==11.1.0
pip==25.0.1
pluggy==1.5.0
prometheus-client==0.21.1
protobuf==5.29.3
pytest==8.3.5
requests==2.32.3
setuptools==75.9.1
six==1.17.0
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
typing-extensions==4.12.2
urllib3==2.3.0
werkzeug==3.1.3
